+++ qunit/tests_attributes.js
@@ -21,4 +21,14 @@ define([
 		$("#testmask").Type("1234.56");
 		assert.equal(testmask.value, "123", "Result " + testmask.value);
 	});
+
+	qunit.test("data-inputmask=\"'mask':'[9-]AAA-999'\" - airomero", function(assert) {
+		var $fixture = $("#qunit-fixture");
+		$fixture.append('<input type="text" id="testmask" data-inputmask="\'mask\':\'[9-]AAA-999\'" />');
+		var testmask = document.getElementById("testmask");
+		Inputmask().mask(testmask);
+
+		$("#testmask").Type("abc123");
+		assert.equal(testmask.value, "ABC-123", "Result " + testmask.value);
+	});
 });