+++ js/inputmask.extensions.js
@@ -37,72 +37,7 @@ Optional extensions on the jquery.inputmask base
 		});
 		Inputmask.extendAliases({
 			"url": {
-				mask: "ir",
-				placeholder: "",
-				separator: "",
-				defaultPrefix: "http://",
-				regex: {
-					urlpre1: new RegExp("[fh]"),
-					urlpre2: new RegExp("(ft|ht)"),
-					urlpre3: new RegExp("(ftp|htt)"),
-					urlpre4: new RegExp("(ftp:|http|ftps)"),
-					urlpre5: new RegExp("(ftp:/|ftps:|http:|https)"),
-					urlpre6: new RegExp("(ftp://|ftps:/|http:/|https:)"),
-					urlpre7: new RegExp("(ftp://|ftps://|http://|https:/)"),
-					urlpre8: new RegExp("(ftp://|ftps://|http://|https://)")
-				},
-				definitions: {
-					"i": {
-						validator: function(chrs, maskset, pos, strict, opts) {
-							return true;
-						},
-						cardinality: 8,
-						prevalidator: (function() {
-							var result = [],
-								prefixLimit = 8;
-							for (var i = 0; i < prefixLimit; i++) {
-								result[i] = (function() {
-									var j = i;
-									return {
-										validator: function(chrs, maskset, pos, strict, opts) {
-											if (opts.regex["urlpre" + (j + 1)]) {
-												var tmp = chrs,
-													k;
-												if (((j + 1) - chrs.length) > 0) {
-													tmp = maskset.buffer.join("").substring(0, ((j + 1) - chrs.length)) + "" + tmp;
-												}
-												var isValid = opts.regex["urlpre" + (j + 1)].test(tmp);
-												if (!strict && !isValid) {
-													pos = pos - j;
-													for (k = 0; k < opts.defaultPrefix.length; k++) {
-														maskset.buffer[pos] = opts.defaultPrefix[k];
-														pos++;
-													}
-													for (k = 0; k < tmp.length - 1; k++) {
-														maskset.buffer[pos] = tmp[k];
-														pos++;
-													}
-													return {
-														"pos": pos
-													};
-												}
-												return isValid;
-											} else {
-												return false;
-											}
-										},
-										cardinality: j
-									};
-								})();
-							}
-							return result;
-						})()
-					},
-					"r": {
-						validator: ".",
-						cardinality: 50
-					}
-				},
+				mask: "(\\http://)|(\\http\\s://)|(ftp://)|(ftp\\s://)*{+}",
 				insertMode: false,
 				autoUnmask: false
 			},