+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file.
 
 ## UNRELEASED
 
+### Updates
+- remove $("selector").inputmask("mask", { mask: "99-999 ..." }) format from plugin
+
 ### Fixed
+- data-inputmask => mask with optionals not parsed correctly #1119
 - Email mask doesn't allow to go to the domain part by mouse #885
 - alias options from 'data-inputmask' is not used anymore #1113
 - Numeric extensions don't supported with vanilla DependencyLib #1116